utils restructured for progress bar and other functions
