+++ brian2modelfitting/modelfitting/__init__.py
@@ -2,15 +2,13 @@
 Package to fit models to experimental data
 """
 
-from .modelfitting import *
 from .modelfittingOOP import *
 from .optimizer import *
 from .metric import *
 from .simulation import *
 from .utils import *
 
-__all__ = ['fit_traces', 'fit_spikes',
-           'generate_fits',
+__all__ = ['callback_setup', 'callback_text', 'make_dic',
            'Optimizer', 'NevergradOptimizer', 'SkoptOptimizer',
            'Simulation', 'RuntimeSimulation', 'CPPStandaloneSimulation',
            'MSEMetric', 'Metric', 'GammaFactor', 'get_gamma_factor', 'firing_rate',
